chore(todo): add custom route for all enpoints
[Finishes #168749165]
