+++ .eslintrc.json
@@ -1,17 +1,28 @@
 {
-  "env": {
-    "es6": true,
-    "node": true,
-    "mocha": true
-  },
-  "extends": ["airbnb-base"],
-  "globals": {
-    "Atomics": "readonly",
-    "SharedArrayBuffer": "readonly"
-  },
-  "parserOptions": {
-    "ecmaVersion": 2018,
-    "sourceType": "module"
-  },
-  "rules": {}
+    "env": {
+        "es6": true,
+        "node": true
+    },
+    "extends": [
+        "airbnb-base"
+    ],
+    "globals": {
+        "Atomics": "readonly",
+        "SharedArrayBuffer": "readonly"
+    },
+    "parserOptions": {
+        "ecmaVersion": 2018,
+        "sourceType": "module"
+    },
+    "rules": {
+        "linebreak-style":"off",
+        "no-console":"off",
+        "camelcase":"off",
+        "import/no-named-as-default":"off",
+        "import/no-named-as-default-member":"off",
+        "consistent-return":"off",
+        "prefer-const":"off",
+        "no-param-reassign":"off",
+        "no-restricted-globals":"off"
+    }
 }
